+++ src/core/objects/empty_placeholder.h
@@ -11,7 +11,7 @@ namespace KyoukoMind
 struct EmptyPlaceHolder
 {
     uint8_t status = DELETED_SECTION;
-    uint32_t bytePositionOfNextEmptyBlock = UNINIT_STATE_32;
+    uint64_t bytePositionOfNextEmptyBlock = UNINIT_STATE_32;
 
 } __attribute__((packed));
